requirements.txt
fastapi
uvicorn[standard]
duckdb
pandas
python-dotenv
langchain
langchain-openai
streamlit
python-multipart